fix build issues.
Still part of moving the snippets
