+++ docs/csharp/language-reference/keywords/snippets/FixedKeywordExamples.cs
