+++ docs/csharp/language-reference/keywords/snippets/GenericWhereConstraints.cs
