+++ docs/csharp/language-reference/keywords/snippets/IterationKeywordsExamples.cs
