+++ docs/csharp/language-reference/keywords/snippets/LockStatementExample.cs
