+++ docs/csharp/language-reference/keywords/snippets/ReadonlyKeywordExamples.cs
@@ -20,20 +20,20 @@ namespace keywords
     // </SnippetReadonlyField>
 
     // <SnippetInitReadonlyField>
-    class SampleClass
+    public class SamplePoint
     {
         public int x;
         // Initialize a readonly field
         public readonly int y = 25;
         public readonly int z;
 
-        public SampleClass()
+        public SamplePoint()
         {
             // Initialize a readonly instance field
             z = 24;
         }
 
-        public SampleClass(int p1, int p2, int p3)
+        public SamplePoint(int p1, int p2, int p3)
         {
             x = p1;
             y = p2;
@@ -42,9 +42,9 @@ namespace keywords
 
         public static void Main()
         {
-            SampleClass p1 = new SampleClass(11, 21, 32);   // OK
+            SamplePoint p1 = new SamplePoint(11, 21, 32);   // OK
             Console.WriteLine($"p1: x={p1.x}, y={p1.y}, z={p1.z}");
-            SampleClass p2 = new SampleClass();
+            SamplePoint p2 = new SamplePoint();
             p2.x = 55;   // OK
             Console.WriteLine($"p2: x={p2.x}, y={p2.y}, z={p2.z}");
         }
@@ -60,13 +60,13 @@ namespace keywords
     {
         public static void Examples()
         {
-            SampleClass.Main();
+            SamplePoint.Main();
             ReadonlyRefReturns();
         }
 
         // <SnippetReadonlyReturn>
-        private static readonly Point origin = new Point(0, 0);
-        public static ref readonly Point Origin => ref origin;
+        private static readonly SamplePoint origin = new SamplePoint(0, 0, 0);
+        public static ref readonly SamplePoint Origin => ref origin;
         // </SnippetReadonlyReturn>
 
         private static void ReadonlyRefReturns()
