+++ docs/csharp/language-reference/keywords/snippets/keywords.csproj
@@ -4,6 +4,7 @@
     <OutputType>Exe</OutputType>
     <TargetFramework>netcoreapp3.1</TargetFramework>
     <Nullable>enable</Nullable>
+    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+    <StartupObject>keywords.Program</StartupObject>
   </PropertyGroup>
-
 </Project>
